noni.

+++ t.py
@@ -4,31 +4,36 @@ from mega import Mega
 import os
 def convert_to_sketch(input_image, output_image):
     try:
-        print(os.listdir()) 
-        # Read the image
-        img = cv2.imread(input_image)
-        if img is None:
-            print("Error: Could not load image.")
-            sys.exit(1)
-        
-        # Convert to grayscale
-        gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
-        
-        # Invert the image
-        inverted = cv2.bitwise_not(gray)
-        
-        # Blur the image
-        blurred = cv2.GaussianBlur(inverted, (21, 21), sigmaX=0, sigmaY=0)
-        
-        # Invert the blurred image
-        inverted_blurred = cv2.bitwise_not(blurred)
-        
-        # Create the pencil sketch
-        sketch = cv2.divide(gray, inverted_blurred, scale=256.0)
-        
-        # Save the output
-        cv2.imwrite(output_image, sketch)
+        if os.path.exists(input_image):
+            print(f"{input_image} file exits...")
+                
+            # Read the image
+            img = cv2.imread(input_image)
+            if img is None:
+                print("Error: Could not load image.")
+                sys.exit(1)
+            
+            # Convert to grayscale
+            gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
+            
+            # Invert the image
+            inverted = cv2.bitwise_not(gray)
+            
+            # Blur the image
+            blurred = cv2.GaussianBlur(inverted, (21, 21), sigmaX=0, sigmaY=0)
+            
+            # Invert the blurred image
+            inverted_blurred = cv2.bitwise_not(blurred)
+            
+            # Create the pencil sketch
+            sketch = cv2.divide(gray, inverted_blurred, scale=256.0)
+            
+            # Save the output
+            cv2.imwrite(output_image, sketch)
+        else:
+            print("img file not exits ",os.listdir())
     except Exception as e:
+        
         print("error : ",e)
 
 if __name__ == "__main__":
